+++ packages/api/src/routes/reports.js
@@ -2,7 +2,7 @@ const express = require('express');
 const fetch = require('node-fetch');
 const _ = require('lodash');
 const config = require('../config/config');
-const { Loo, Report } = require('@neontribe/gbptm-loodb')(config.mongo.url);
+const { Loo, Report } = require('../db')(config.mongo.url);
 const { checkJwt, checkScopes } = require('../config/auth');
 
 const router = express.Router();