+++ packages/ui/src/pages/AddEditPage.js
@@ -20,7 +20,7 @@ import {
 import { actionHighlight } from '../redux/modules/mapControls';
 
 import config from '../config';
-import { mappings } from '@neontribe/api-client';
+import { mappings } from '@toiletmap/api-client';
 
 import styles from './css/edit-loo-page.module.css';
 import layout from '../components/css/layout.module.css';