+++ packages/leaflet-control-geocoder/CHANGELOG.md
@@ -5,7 +5,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # [2.8.0](https://github.com/perliedman/leaflet-control-geocoder/compare/v2.7.0...v2.8.0) (2018-10-11)
 
-**Note:** Version bump only for package @neontribe/leaflet-control-geocoder
+**Note:** Version bump only for package @toiletmap/leaflet-control-geocoder