+++ packages/leaflet-control-geocoder/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0](https://github.com/perliedman/leaflet-control-geocoder/compare/v2.9.0...v3.0.0) (2019-10-14)
+
+
+### Bug Fixes
+
+* dependency updates ([3f75a8c](https://github.com/perliedman/leaflet-control-geocoder/commit/3f75a8c))
+* deps ([c5c3b71](https://github.com/perliedman/leaflet-control-geocoder/commit/c5c3b71))
+* deps ([8e60716](https://github.com/perliedman/leaflet-control-geocoder/commit/8e60716))
+* update dependencies ([c89240f](https://github.com/perliedman/leaflet-control-geocoder/commit/c89240f))
+
+
+### Features
+
+* move to new data structures ([ab17e0e](https://github.com/perliedman/leaflet-control-geocoder/commit/ab17e0e))
+
+
+### BREAKING CHANGES
+
+* requires full database migration, see the migrate
+package
+
+
+
+
+
 # [2.8.0](https://github.com/perliedman/leaflet-control-geocoder/compare/v2.7.0...v2.8.0) (2018-10-11)
 
 **Note:** Version bump only for package @toiletmap/leaflet-control-geocoder